Rename llumnix untils to llumnix entrypoints
