+++ llumnix/entrypoints/vllm/api_server.py
@@ -28,11 +28,14 @@ from vllm.engine.arg_utils import AsyncEngineArgs
 from vllm.engine.async_llm_engine import AsyncStream
 
 from llumnix.arg_utils import EngineManagerArgs
-from llumnix.entrypoints.llumnix_utils import (get_ip_address,
-                                               launch_ray_cluster, connect_to_ray_cluster,
-                                               is_gpu_available, init_llumnix_components,
-                                               LlumnixArgumentParser,
-                                               init_per_token_latency_breakdown_dict, record_per_token_latency_breakdown)
+from llumnix.entrypoints.llumnix_entrypoints import (get_ip_address,
+                                                     launch_ray_cluster,
+                                                     connect_to_ray_cluster,
+                                                     is_gpu_available,
+                                                     init_llumnix_components,
+                                                     LlumnixArgumentParser,
+                                                     init_per_token_latency_breakdown_dict,
+                                                     record_per_token_latency_breakdown)
 from llumnix.logger import init_logger
 from llumnix.utils import random_uuid
 from llumnix.backends.vllm.utils import check_engine_args
